requirements.txt
spikingjelly==0.0.0.0.14
tensorboard==2.19.0
scipy==1.15.3
wandb
pandas
tabulate
pyyaml